src/requirements.txt
alpaca-trade-api 
python-dotenv 
pandas 
pytz 
matplotlib
backtrader